requirements.txt
onnx
sanic
pytest
pillow
sanic-testing
onnxruntime
opencv-python-headless